python/requirements.txt
dateparser==1.1.8
mysql-connector-python==8.0.33
pydantic==1.10.7
black==23.7.0
